+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to
 
 - Isolate failed to refresh token errors from other oauth errros
   [#3332](https://github.com/OpenFn/lightning/issues/3332)
+- Handling of CLI error messages when fetching metadata via an adaptor
+  [#3367](https://github.com/OpenFn/lightning/issues/3367)
 
 ## [v2.13.4] - 2025-07-05